feat: split subscriber + dockerfile
