+++ Dockerfile
@@ -1,6 +1,7 @@
 FROM node:gallium-alpine
 ARG BUILD_CONTEXT
-ENV RUN_CONTEXT ${BUILD_CONTEXT}
+ARG RUN_COMMAND
+ENV RUN_CONTEXT ${BUILD_CONTEXT} ${RUN_COMMAND}
 
 WORKDIR /usr/src/app
