+++ packages/application/src/subscribers/index.ts
@@ -3,15 +3,12 @@ dotenv.config();
 
 import 'reflect-metadata';
 
-import { initialise } from './startup';
-import { TYPES } from './types';
-
-// eslint-disable-next-line import/order
+import { initialise } from '../startup';
+import { TYPES } from '../types';
 import { IEventBus } from '@cqrs-es/core';
 
 (async () => {
   const container = await initialise();
-
   const baseEventHandler = container.get<IEventBus>(TYPES.EventBus);
   baseEventHandler.subscribeEvents();
 })();
